Added README / Updated transcript generation
Attempt at implementing default transcript generation from EAF files.

NOTE: This feature is not suited for general ELAN projects. This is an attempt to automatically generate transcripts from a specific formatting of tiers.

In the future, there should really be a way to specify which tiers should be included in the transcript, just like you do in ELAN.